+++ docs/changelog.rst
@@ -12,6 +12,7 @@ Changelog
 * Added ``make bump`` target and replaced ``make env`` with ``make init`` for
   setting up the development environment.
 * Removed deprecated ``hl7.version`` module; use ``hl7.__version__`` instead.
+* Fixed a bug in `mllp_send --quiet` -- thanks `Spencer Vecile <https://github.com/svecile>`_!
 * Switched formatting and linting from black/isort and flake8 to ``ruff``.